+++ Assassin/TransferTask.cpp
@@ -6,10 +6,22 @@ using namespace Assassin;
 using namespace Platform;
 using namespace Windows::Networking::BackgroundTransfer;
 
+extern ULONGLONG M2GetTickCount();
+
+void Assassin::TransferTask::RaisePropertyChanged(String ^ PropertyName)
+{
+	using Windows::UI::Xaml::Data::PropertyChangedEventArgs;
+	this->PropertyChanged(
+		this, ref new PropertyChangedEventArgs(PropertyName));
+}
+
 TransferTask::TransferTask(DownloadOperation^ Operations)
 	: m_Operation(Operations)
 {
-	
+	if (TransferTaskStatus::Running == this->Status)
+	{
+		this->m_Operation->AttachAsync();
+	}
 }
 
 Uri^ TransferTask::RequestedUri::get()
@@ -55,6 +67,39 @@ uint64 TransferTask::BytesReceived::get()
 	return this->m_Operation->Progress.BytesReceived;
 }
 
+uint64 TransferTask::BytesReceivedSpeed::get()
+{	
+	if (0 != this->m_LastUpdated)
+	{	
+		uint64 DeltaBytesReceived =
+			this->BytesReceived - this->m_LastBytesReceived;
+		ULONGLONG DeltaPassedTime = 
+			M2GetTickCount() - this->m_LastUpdated;
+
+		this->m_BytesReceivedSpeed = 
+			DeltaBytesReceived * 1000 / DeltaPassedTime;
+	}
+	
+	this->m_LastUpdated = M2GetTickCount();
+	this->m_LastBytesReceived = this->BytesReceived;
+
+	return this->m_BytesReceivedSpeed;
+}
+
+uint64 TransferTask::RemainTime::get()
+{
+	uint64 ReceivedSpeed = this->m_BytesReceivedSpeed;
+	
+	if (0 == ReceivedSpeed)
+	{
+		return static_cast<uint64>(-1);
+	}
+	else
+	{
+		return (this->TotalBytesToReceive - this->BytesReceived) / ReceivedSpeed;
+	}
+}
+
 uint64 TransferTask::TotalBytesToReceive::get()
 {
 	return this->m_Operation->Progress.TotalBytesToReceive;
@@ -62,15 +107,32 @@ uint64 TransferTask::TotalBytesToReceive::get()
 
 void TransferTask::Pause()
 {
-	this->m_Operation->Pause();
+	if (TransferTaskStatus::Running == this->Status)
+	{
+		this->m_Operation->Pause();
+	}
 }
 
 void TransferTask::Resume()
 {
-	this->m_Operation->Resume();
+	if (TransferTaskStatus::Paused == this->Status)
+	{
+		this->m_Operation->Resume();
+	}
 }
 
 void TransferTask::Cancel()
 {
-	this->m_Operation->AttachAsync()->Cancel();
+	if (TransferTaskStatus::Canceled != this->Status)
+	{
+		this->m_Operation->AttachAsync()->Cancel();
+	}
+}
+
+void Assassin::TransferTask::NotifyPropertyChanged()
+{
+	this->RaisePropertyChanged(L"BytesReceived");
+	this->RaisePropertyChanged(L"BytesReceivedSpeed");
+	this->RaisePropertyChanged(L"RemainTime");
+	this->RaisePropertyChanged(L"TotalBytesToReceive");
 }